API_Handler/requirements.txt
flask
pandas
sklearn
python-dotenv
couchdb
flask-cors
requests